virtual-production/wearable-rpi/requirements.txt
# Core dependencies for Raspberry Pi wearable sensor system

# Sense HAT library
sense-hat>=2.4.0

# Camera support (use legacy camera interface for better compatibility)
picamera2>=0.3.16
opencv-python>=4.8.0

# Audio processing
pyaudio>=0.2.11
numpy>=1.24.0
scipy>=1.10.0

# API communication
requests>=2.31.0
aiohttp>=3.8.0
websocket-client>=1.6.0  # VP WebSocket 클라이언트

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging
colorlog>=6.7.0

# Data processing
pandas>=2.0.0

# Threading and async support
asyncio-mqtt>=0.16.0  # Optional: for MQTT support

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0